Bug fixes, 8 new features
Bug fixes, 8 new features, this is the first proof of concept

+++ kernel/CDelegates.java
@@ -7,18 +7,19 @@ public class CDelegates
 {
     public CDelegates() throws Exception
     {
-        ResultSet rs=UTILS.DB.executeQuery("SELECT COUNT(*) AS total from blocks");
+        ResultSet rs=UTILS.DB.executeQuery("SELECT COUNT(*) AS total FROM blocks");
         rs.next();
-        this.refresh(0);
+        
+        if (rs.getLong("total")==1) 
+            this.refresh(0);
     }
     
     public long computePower(String delegate) throws Exception
     {
-       ResultSet rs=UTILS.DB.executeQuery("SELECT SUM(adr.balance) AS total "
-                                          + "FROM del_votes AS dv "
-                                          + "JOIN adr ON adr.adr=dv.adr "
-                                         + "WHERE dv.delegate='"+delegate+"' "
-                                           + "AND dv.type='ID_UP'");
+       ResultSet rs=UTILS.DB.executeQuery("SELECT SUM(power) AS total "
+                                          + "FROM del_votes "
+                                         + "WHERE delegate='"+delegate+"' "
+                                           + "AND type='ID_UP'");
        
        // Next
        rs.next();
@@ -26,11 +27,10 @@ public class CDelegates
        // Upvotes power
        long up=Math.round(rs.getDouble("total"));
        
-       rs=UTILS.DB.executeQuery("SELECT SUM(adr.balance) AS total "
-                                          + "FROM del_votes AS dv "
-                                          + "JOIN adr ON adr.adr=dv.adr "
-                                         + "WHERE dv.delegate='"+delegate+"' "
-                                           + "AND dv.type='ID_DOWN'");
+       rs=UTILS.DB.executeQuery("SELECT SUM(power) AS total "
+                                + "FROM del_votes "
+                               + "WHERE delegate='"+delegate+"' "
+                                 + "AND type='ID_DOWN'");
        
         // Next
        rs.next();
@@ -42,8 +42,43 @@ public class CDelegates
        return (up-down);
     }
     
+    public void refreshVotesPower() throws Exception
+    {
+        // Load votes    
+        ResultSet rs=UTILS.DB.executeQuery("SELECT * "
+                                           + "FROM del_votes AS dv "
+                                           + "JOIN adr ON adr.adr=dv.adr");
+        
+        while (rs.next())
+        UTILS.DB.executeUpdate("UPDATE del_votes "
+                                + "SET power='"+Math.round(rs.getDouble("balance"))+"' "
+                              + "WHERE ID='"+rs.getLong("ID")+"'");
+    }
+    
+    public void logDelegates(long block) throws Exception
+    {
+        // Remove delegates with the same block
+        UTILS.DB.executeUpdate("DELETE FROM delegates_log "
+                                  + "WHERE block='"+block+"'");
+        
+        // Load delegates 
+        ResultSet rs=UTILS.DB.executeQuery("SELECT * FROM delegates");
+        
+        while (rs.next())
+           UTILS.DB.executeUpdate("INSERT INTO delegates_log "
+                                        + "SET delegate='"+rs.getString("delegate")+"', "
+                                            + "power='"+rs.getLong("power")+"', "
+                                            + "block='"+rs.getLong("block")+"'");
+    }
+    
     public void refresh(long block) throws Exception
     {
+        // Block
+        if (block%100!=0) return;
+            
+        // Votes power
+        this.refreshVotesPower();
+        
         // Delete delegates
         UTILS.DB.executeUpdate("DELETE FROM delegates");
         
@@ -58,32 +93,24 @@ public class CDelegates
             // Insert
             UTILS.DB.executeUpdate("INSERT INTO delegates "
                                          + "SET delegate='"+rs.getString("delegate")+"', "
-                                             + "power='"+power+"'");
+                                             + "power='"+power+"', "
+                                             + "block='"+block+"'");
         }
+        
+        // Load delegates                                     
+        rs=UTILS.DB.executeQuery("SELECT * "
+                                 + "FROM delegates");
+        
+        // Parse
+        while (rs.next())
+            UTILS.DB.executeUpdate("UPDATE delegates "
+                                    + "SET dif='"+UTILS.BASIC.formatDif(this.getDif(rs.getString("delegate")).toString(16))+"'");
         
         // Delete downvoted delegates
-        UTILS.DB.executeUpdate("DELETE FROM delegates WHERE power<1");
-        
-        // Only the first 100
-        rs=UTILS.DB.executeQuery("SELECT * "
-                                 + "FROM delegates "
-                             + "ORDER BY power DESC "
-                                + "LIMIT 0,100");
-        
-        double min=0;
-        long del_no=0;
-        while (rs.next()) 
-        {
-            del_no++;
-            min=rs.getDouble("power");
-        }
-        
-        // Removes 
-        if (del_no>100)
-           UTILS.DB.executeUpdate("DELETE FROM delegates WHERE power<="+min);
-        
-        // Power less than 1
         UTILS.DB.executeUpdate("DELETE FROM delegates WHERE power<10");
+        
+        // Log 
+        this.logDelegates(block);
     }
     
     public long getPower(String delegate) throws Exception
@@ -94,13 +121,25 @@ public class CDelegates
         
         // No delegate
         if (!UTILS.DB.hasData(rs))
-            throw new Exception("Invalid delegate");
+            return 1;
             
         // Next
         rs.next();
         
         // Power
         return rs.getLong("power");
+    }
+    
+    public long getTotalPower() throws Exception
+    {
+        ResultSet rs=UTILS.DB.executeQuery("SELECT SUM(power) AS total "
+                                           + "FROM delegates");
+        
+        // Next
+        rs.next();
+        
+        // Power
+        return rs.getLong("total");
     }
     
     public boolean isDelegate(String adr) throws Exception
@@ -115,25 +154,49 @@ public class CDelegates
             return false;
     }
     
-    public boolean canMine(String adr) throws Exception
-    {
-        ResultSet rs=UTILS.DB.executeQuery("SELECT * "
-                                           + "FROM delegates "
-                                          + "WHERE delegate='"+adr+"'");
-        
-        if (UTILS.DB.hasData(rs)==false) return false;
-        
-        // Next 
-        rs.next();
-        
-        if (rs.getLong("mined_24")<=rs.getLong("max_blocks"))
-            return true;
-        else
-            return false;
-    }
+    
     
     public BigInteger getDif(String delegate) throws Exception
     {
         return UTILS.NET_STAT.net_dif.multiply(BigInteger.valueOf(this.getPower(delegate)));
+    }
+    
+    public long getMaxBlocks(String delegate) throws Exception
+    {
+        // Power
+        long power=this.getPower(delegate);
+        
+        // Total power
+        long total_power=this.getTotalPower();
+        
+        // Percent
+        double p=power*100/total_power;
+        
+        // Blocks
+        long blocks=Math.round(p/100*1440)*5;
+        
+        // Return
+        return blocks;
+    }
+    
+    public long getMinedBlocks(String delegate, long block) throws Exception
+    {
+        ResultSet rs=UTILS.DB.executeQuery("SELECT COUNT(*) AS mined "
+                                           + "FROM blocks "
+                                          + "WHERE block<"+(block-1440));
+        
+        // Next
+        rs.next();
+        
+        // Return
+        return rs.getLong("mined");
+    }
+    
+    public boolean canMine(String delegate, long block) throws Exception
+    {
+        if (this.getMinedBlocks(delegate, block)<this.getMaxBlocks(delegate))
+            return true;
+        else
+            return false;            
     }
 }
